MAINT: Move testlayer to `testing` module, like all vivi packages do
