+++ src/zeit/find/browser/tests/test_doctest.py
@@ -1,9 +1,9 @@
 import zeit.cms.testing
-import zeit.find.tests
+import zeit.find.testing
 
 
 def test_suite():
     return zeit.cms.testing.FunctionalDocFileSuite(
         'README.txt',
         package='zeit.find.browser',
-        layer=zeit.find.tests.LAYER)
+        layer=zeit.find.testing.LAYER)
