+++ src/zeit/find/tests/test_search.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 from zope.component import getUtility
 from zeit.cms.testing import FunctionalTestCase
 from zeit.find.elastic import ICMSSearch
-from zeit.find.tests import LAYER
+from zeit.find.testing import LAYER
 
 
 class TestElasticsearch(TestCase):
